Add necessary implementation to pass all BasicsTests
Added implementation of cass_value_get_bytes to retrieve bytes of
a Blob value.

Added paging state token setter for CassStatement.

Added paging state token getter from query result.

The database version is set to 3.0.8 if the provided version in tests
has prefix `release:`. This will skip some of the tests which are also
skipped in the cpp-driver tests and are not disabled.