Change CassValue to contain full information about collections
Previously CassValue contained value of type Option<CqlValue> which did
not allow to retrieve type information for collection elements,
particularly if the collection is empty its type will be unknown.

A new enum Value is introduced which separates collection and
non-collection types and allows to store type information for collection
elements.