Add collection iterator implementation
Added CassCollectionIterator struct which keeps reference to the
collection and the position of next element in the collection to be
consumed.

Added implementation of cass_iterator_get_value for List, Set and Tuple
collections.

Changed value type getter to return CASS_VALUE_TYPE_TEXT for both
Varchar and Text types as Varchar is not supported by rust driver.

As the CassandraTypesTests for collections were dependent on the C++
driver implmenetation, the tests were modified to get `CassValue` from
the query result, validate that the value is a collection, and then
iterate over its elements and validate them against the expected values.

e.g.
  Tuple result_tuple(result.first_row().next().as<Tuple>());

Here, the `next()` call returns a pointer to a CassValue defined in the
rust bindings, so casting it to a Tuple defined in the C++ driver is an
invalid operation. Instead, one should use the functions defined in
cassandra.h to get access to the elements of the collection:

e.g.
  * cass_iterator_from_tuple(tuple)
  * cass_value_item_count(tuple)