Add materialized view metadata
Added `CassMaterializedViewMeta` struct to store metadata for views. As
the interface should enable to get base table metadata for each view, as
well as to get all views for a given table metadata, to avoid memory
leaks because of the cyclic references, `base_table` of
`CassMaterializedViewMeta` is a `Weak` reference to `CassTableMeta`.