Avoid creating references to uninitialized data
Creating a reference to uninitialized data is most of the time
undefined behaviour in Rust - even if data under this reference is
never read.

This commit removes all occurences of this problem (in accordance with
cpp-driver's semantics), instead using `std::ptr::write` - safer and
more explicit way to achieve the same goal.