Increment Cassandra version in CI tests
The `TracingTests.Simple` test was flaky, supposedly because of
Cassandra version 3.0.8. A select all query from
`system_traces.sessions` sometimes returned empty result, however,
a manual check showed that the table was not empty during the execution
of the query. The issue is fixed in 3.0.9 which waits for the tracing
events before returning response and queries at same consistency level
client side.

Ref to the changelog: https://raw.githubusercontent.com/apache/cassandra/cassandra-3.0.9/CHANGES.txt

The version is incremented to 3.0.9 for now, as it allows to
avoid the issue and to not introduce new tests which were skipped
previously because of the version.