Add install capabilities to top-level CMake config
Before this change, the generated Makefile by CMake lacked the 
capability to run "make install". This patch makes it possible
to run "make install", for example:

mkdir build
cd build
cmake -DCMAKE_BUILD_TYPE=RELEASE -DCASS_INSTALL_PKG_CONFIG=OFF -DCMAKE_INSTALL_PREFIX:PATH=/usr -DCMAKE_INSTALL_LIBDIR=/usr/lib64 ..
make DESTDIR=/tmp/example install

The names and contents of installed files try to mimic the original
C++ Driver files. For example, installation of C++ Driver generates the 
following .so files/symlinks:

usr/lib64/libscylla-cpp-driver.so -> libscylla-cpp-driver.so.2
usr/lib64/libscylla-cpp-driver.so.2 -> libscylla-cpp-driver.so.2.16.2
usr/lib64/libscylla-cpp-driver.so.2.16.2

This particular behavior is replicated in a new build-cmake-compat.sh
script, which is executed as a BUILD_COMMAND in the Rust external 
project in the top-level CMakeLists.txt.

Unfortunately, I have not been able to configure CMake to create
such symlinks on its own. One of the reasons is an unfixed bug in CMake:
https://gitlab.kitware.com/cmake/cmake/-/issues/23249#note_1277271.
Therefore, the build-cmake-compat.sh is introduced to create the
necessary files in a way it's easy to install for CMake.