Include static library in installed files
Add copying of .a static library file to build-cmake-compat.sh file
when CASS_BUILD_STATIC is ON. The copied file name mimics the filename
of the original C++ driver.

After this commit, the contents of DESTDIR (as in make DESTDIR=... install)
are the same between the C++-Rust bindings and the original C++ driver
(apart from the usr/include/dse.h file).