Session: Store name -> exec profile mapping
The mapping is cloned from the CassCluster to the CassSession on
`cass_session_connect()`. Additionally, while the mapping in Cluster is
from name into profile builders, the mapping in Session is into already
built profile handles.
Note that `ExecutionProfileBuilder::build()` may not be called outside
async context, because building DefaultPolicy panics outside it
(most likely because of `tokio::spawn()` usage
in `LatencyAwarenessBuilder::new()`).