CassFuture: make_raw() returns *mut
The return type of make_raw() was erroneously set to *const
CassFuture, which both is incompatible with the Cpp driver API in
cassandra.h (which uses C non-const *) and leads to inconvenient usage
due to mandatory use of casts from *const to *mut (e.g. in case of
`cass_future_error_message()`).