session: Regression test Cluster use-after-free
It is an annoying bug to refer to CassCluster object after is has been
freed. This may happen in the future returned by cass_session_connect().
The test runs a subprocess (thanks to rusty_fork crate) which segfaults
if the cluster is referenced by the future, so then the test clearly
fails.