Cargo.toml: Update rust driver version
The new deserialization api in the Rust driver is still a WIP, so the
version for the moment references to my fork. The `scylla-proxy` crate
version also relies on my repo to avoid version conflicts.
Added error conversion for `ParseError` into `CassError`, albeit we will
need to have correct and more proper mappings.
For quick migration, legacy query result type is used.