Refactor collection iterators to support lazy deserialization
Make `cass_value_is_collection` independent from value field of
`CassValue` as the field is redundant after the addition of lazy
deserialization.
The same is done for `cass_value_item_count` and all collection
iterators which are dependent on that field.
Prepared collection iterator structs to add lazy deserialization of
collection values.