cass_prepared: hold info about col data types
Until now, we would not hold the information about column data types
from PreparedMetadata. We need to hold this information to perform
a typecheck during binding values to statement.

We could construct the data types from column specs each time
we bind a value. However, CassDataType might be a heavy nested
object, and so I decided to cache it in CassPrepared.