ser: use rust builtin serialization for simple types
Previously, we would reimplement a serialization logic. It was a copy
from rust driver, with the typechecks excluded.

In this commit, we get rid of the repeated code. To skip the
typechecks (or rather, to make the typechecks always pass), we simply
provide a valid ColumnType to SerializeValue::serialize().

This is a kind of hack, but thanks to that we don't have to reimplement
serialization logic.