test_collection: adjust checking collection emptiness
`cass_iterator_get_value` function should not be used to
determine whether the collection is empty... It was working for
cpp-driver, since this function always returns a non-null pointer there.
In case of cpp-rust-driver, we return a null pointer when user calls
`cass_iterator_get_value` before `cass_iterator_next`, because the latter
function is responsible for advancing the iterator, whose state is
uninitialized upon creation (position is None). OTOH, cpp-driver
returns a pointer to uninitialized value when calling `cass_iterator_get_value`
first - this is why `is_null_` was always false for cpp-driver.

Now, to check if collection is empty, we make use of `cass_value_item_count`
API function.