test_collection: set is_null_ to false when created from non_null value
From what I understood, the `is_null_` flag indicates whether
collection value is null. For example, this flag is used to decide
if we should call `cass_tuple_set_null` or `cass_tuple_set_collection`.

In Collection::initialize, we already check that value is non-null.
I believe this was a bug in the original tests. The fact that collection
is empty, does not mean that it is null.

In addition, `cass_iterator_get_value` function should not be used to
determine whether the collection is empty... It was working for
cpp-driver, since this function always returns a non-null pointer.
In case of cpp-rust-driver, we return a null pointer when user calls
`cass_iterator_get_value` before `cass_iterator_next`, because the latter
function is responsible for advancing the iterator, whose state is
uninitialized upon creation (position is None). OTOH, cpp-driver
returns a pointer to uninitialized value when calling `cass_iterator_get_value`
first - this is why `is_null_` was always false for cpp-driver.