cargo: update rust edition to 2021
This is required to make use of CStr literals - null terminated
string literals, e.g. c"foobarbaz".

We needed to adjust a code in one place. Starting from edition 2021,
closures capture only the object's fields that are actually used (instead
of captuting whole object). In our case, we would capture underlying
pointer, which is not `Send` (instead of a whole wrapper, which
has an unsafe Send implementation).