integration: fix SerialConsistencyTests
Original cpp-driver allows user to SET the invalid serial consistency.
Then, the request is sent with an invalid serial consistency, resulting in a server error.
However, rust-driver comes with a type safety in this matter, disallowing
the user to set invalid serial consistency, thus we are not able to mimic this behaviour in cpp-rust-driver.
We need to adjust the integration test, and assert that setting the invalid serial consistency fails with an error.