errors: map UnknownNamedParameterError to CassError
Unfortunately, there is no other way to inspect what's
inside the SerializationError (since its only member is private),
other than just converting it to string and checking string contents.

We probably need to make this member public in rust-driver (so users
can downcast to their custom errors).

In this commit, we also adjust one of the NamedParameterTests to this change.
We are going to enable this test suite in the next commit.