statement: extract common unquoting logic
Previously, we would store a quoted parameter name for simple queries.
This is a bug.
The quotes are only introduces to decide whether the name
is case sensitive for prepared statements.

To fix this, I extracted the common logic of unquoting to
`CassStatement::bind_cql_value_by_name`. It will pass the unquoted
parameter name to methods on corresponding statement kind.