build: generate definition for size_t
In this commit we tell bindgen to generate a binding for size_t type.
Actually, we only tell it not to treat `size_t` as `usize` (which is
done by default).

Why there were no clashes with `size_t` definition, even though
we used `.allowlist("size_t")`?
The docstring of `.size_t_is_usize()` states:
```
Set whether size_t should be translated to usize.

If size_t is translated to usize, type definitions for size_t will not be emitted.

size_t is translated to usize by default.
```

This means that previously (by default), definition for `size_t` was not
generated in `basic_types.rs`. What happens now is the definition is
generated based on the architecture.
In my case it's std::os::raw::c_ulong (equivalent to u64).

I'm not really sure whether we should treat `size_t` as `usize`
(i.e. create an alias `type size_t = usize`) or not (i.e. `.size_t_is_usize(false)`).
This is my question to reviewers :).