build: use cassandra.h from top-level include directory
Currently, we use two cassandra.h files during build.
- include/cassandra.h -> an original file, copied from cpp-driver
- scylla-rust-wrapper/extern/cassandra.h -> a copy of the above file,
  used to generate bindings in build.rs.

Having both of the files is error-prone, since we need to remember
to update both of the files in case of some changes. We also need to
make sure that their contents are the same.

In this commit, I update `build.rs` to point to original cassandra.h
file (which is out of the scope of Cargo project, but it does not
seem to be a problem).