result: replace col_specs with col_names
Currently, we heavily depend on the fact that `QueryResult::col_specs()`
returns a  `&[ColumnSpec<'static>]`. This won't be true once we bump
rust-driver dependency to 0.15 - the lifetime of ColumnSpec will be tied
to the lifetime of `QueryRowsResult` object.

It looks like most of the information held by `col_spec` was reduntant.
It was only used to retrieve the name of the column. This is why, we will
only hold the information about the column names. The column types are
stored in `col_data_types`.

This also allows us to get rid of `'static` bound from `CassResultData::from_result_payload`.
It will be helpful when bumping to 0.15.

As a bonus, we can also replace the `Vec<>` with a slice `&[]` in
`from_result_payload`, since we do not need to own the column specs anymore.