tests: disable AlterProperlyUpdatesColumnCount
This test should not be enabled!

See the description:
 * Verify that the column count of a bound statement's result metadata is
 * properly updated for newer protocol versions (v5 and greater) when a table's
 * schema is altered.

This test will not work after next commit, thus it's being disabled.

On the other hand, the test AlterDoesntUpdateColumnCount should be enabled.
 * Verify that the column count of a bound statement's result metadata doesn't
 * change for older protocol versions (v4 and less) when a table's schema is altered.

It will be enabled later in this PR, once following issue is addressed:
Currently, the `column_specs` and `col_data_types` may have different sizes
for prepared statements. This is because, `col_data_types` is created based
on prepared statement's cached metadata, and `column_specs` are taken from
the QueryResult.