metadata: introduce CassColumnSpec
This structure will hold information about the column name
and the column data type.

Refactored the `CassResultData`, so now it holds one vector
of `CassColumnSpec` instead of two separate vectors for names and
data types. Previous version was buggy for prepared statements:
- data types came from cached metadata from the statement
- column names came from metadata provided by query result

Because of this, the vector could be of a different length, when someone
ALTERed the corresponding table.

Notice that after this commit, we don't reuse cached metadata from prepared
statement and reuse the metadata provided from the server. Next commits
will fix that.