cargo: bump rust-driver to 0.15
1. CassResult was refactored yet again.
Now, we clearly distinguish betwen Rows and non-Rows
result. `CassResultKind` and `CassRowsResult` are
introduced. `CassRowsResult` holds information about
the (for now, eagerly deserialized) rows and metadata.
The usages of `CassResult` are adjusted throughout the codebase.

2. CassErrorResult was extended by two variants:
- metadata deserialization error - happens during conversion from QueryResult
  to QueryRowsResult
- deserialization error - happens during eager rows deserialization.

3. CassResult construction logic was adjusted to new QueryResult.
As mentioned above, some errors can occur during conversion, thus
`CassResult::from_result_payload` returns a StdResult now.
For now, we eagerly deserialize the rows. This is equivalent to the
version before this PR.