cargo: add thiserror dependency
Once we bump to 0.15, the error can occur during the conversion to
QueryRowsResult. The returned error is not a QueryError, but a standalone
type. This is why, we need to extend our `CassErrorResult` so it's an
enum over possible error types returned during query execution and conversions.

thiserror crate will come in handy, when defining such error type. It allows
to quickly generate `From` implementations for corresponding variants
and most importantly, it helps generating `std::fmt::Display` implementation.