collection: allocate static empty collections
Other parts of the code make an assumption, that the pointer
representing `CassDataType` was obtained from an Arc allocation.

Take for example `cass_data_type_add_sub_type` - it clones an Arc.

This is a bug, that was fortunately detected by applying more restrictions
on the pointer types (introduced later in this PR).