metadata: try to upgrade Weak before returning a pointer
Weak::as_ptr() can return an invalid pointer. It can be even dangling (non-null).

It's safer to try to upgrade to an Arc. If upgrade was successful,
make use of RefFFI API to return a valid pointer. Otherwise, return
non-dangling null pointer.

In previous PR, there was a suggestion to introduce a test case for this.
However, if this upgrade returns `None`, this indicates a serious bug in the driver. This is because
both `CassTableMeta` and `CassMeterializedViewMeta` are owned by `CassSchemaMeta`.
The upgrade would return None only if someone freed `CassSchemaMeta` (cass_schema_meta_free).
But in such case, the corresponding CassMaterializedViewMeta was freed as well.