future: pass future pointer to callback directly
Before this commit, the pointer was obtained from a valid
reference &CassFuture, which is totally fine.

However, I want to reduce the ways one can obtain such pointer.
For ArcFFI (shared pointers), I want them to be obtainable only in two ways:
- `ArcFFI::as_ptr()` which accepts an &Arc
- from the user, as a function parameter

This way, we are guaranteed that the pointer comes from a valid Arc allocation
(unless user provided pointer to some garbage, but there is no much we can do about it).
If we assume that user provides a pointer returned from some prior call to
API, we are guaranteed that it is valid, and comes from an Arc allocation (or is null).

I don't want to allow ArcFFI api to create a pointer from a refernce, to prevent
creating a pointer, from example from stack allocated object:
```
let future = CassFuture { ... };
let future_ptr = ArcFFI::as_ptr(&future);
```

Since some methods started accepting a pointer, they are marked as unsafe, because
they assume that pointer is valid. We will be able to remove unsafes once
CassPtr is introduced - the type will give enough guarantees for these methods
to be safe.

In previous PR, there was a question regarding CassFuture::set_callback:

Q: Here self_ptr should refer to the same CassFuture as self, right?
Can you instead make the receiver of this function an arc? self: &Arc<CassFuture>

A:
I don't see a good reason behind it. We get a pointer in cass_future_set_callback,
then we would need to convert it to Arc (increase a refcount), just to then convert it
back to pointer (decrease refcount) and pass the ponter to callback.