prepared: enable cached result metadata optimization by default
We already cache the metadata on cpp-rust-driver side (see CassPrepared::result_metadata field),
thus we can enable the optimization on rust-driver side as well. This will prevent the server
from sending redundant bytes representing a result metadata during EXECUTE.

NOTE: We are aware that it makes cached metadata immutable. It is expected, though - there
is an integration test for this for CQL protocol v4 (AlterDoesntUpdateColumnCount).
This issue is addressed in CQL protocol v5, but Scylla doesn't support it yet, and probably
won't support it in the near future.