bound: move bound_values to lower layer
Current binding code is fairly complicated. This is because we need
to match against the &[mut] self.statement in each function.

To prevent this, we move the bound_values to lower layer - i.e.
BoundPreparedStatement and BoundSimpleQuery.

This commit does not change the logic at all. It is equivalent to the
previous version.