result: introduce CassRawValue and CassRawRow
CassRawValue:
It is a simple wrapper over the ColumnType and optional FrameSlice. In addition,
it contains information about the "item_count" which tells how many sub-values
it contains (applicable for collections, tuples and UDTs).

The `SerializeValue` implementation for CassRawValue handles the "empty" values
accordingly (the same way cpp-driver does).

CassRawRow:
It is a simple wrapper over Vec<CassRawValue> which implements the `SerializeRow` trait.
We deserialize directly to `Vec<_>`, because there are methods that request a column
from row at given index. This is what cpp-driver does as well.