session: pass Arc to async block instead of reference
cpp-driver assumes that session object can be prematurely dropped. This means,
that we should increase the reference count in functions where we pass the session
to an async block. This will prevent UAF. There actually is a test case
for this (AsyncTests::Close). However, we cannot enable it yet,
since it expects that prematurely dropped session awaits all async
tasks and before closing.