cass_types: replace From<NativeType> for CassValueType
From<> impl takes an owned NativeType. Because of that, we need to
clone the NativeType when we construct CassValueType (because we only
have access to the reference).

As of now, the clone() is not a problem, because NativeType could easily
derive Copy. It may not be the case in the future, so let's protect ourselves
from implicit, potentially expensive cloning.