metadata: replace get_column_type_from_cql_type() usage
Since CqlType is no longer there, we can reuse get_column_type() method
which is used for conversion from ColumnType.

Notice that get_column_type_from_cql_type() accepted two additional arguments:
- keyspace UDTs (map)
- keyspace name

It was "used" in only one place:
```
    pub fn create_with_params(
        user_defined_types: &HashMap<String, Arc<UserDefinedType>>,
        keyspace_name: &str,
        name: &str,
        frozen: bool,
    ) -> UDTDataType {
        UDTDataType {
            field_types: user_defined_types
                .get(name)
                .map(|udf| &udf.field_types)
                .unwrap_or(&vec![])
                .iter()
                .map(|(udt_field_name, udt_field_type)| {
                    (
                        udt_field_name.clone().into_owned(),
                        Arc::new(get_column_type_from_cql_type(
                            udt_field_type,
                            user_defined_types,
                            keyspace_name,
                        )),
                    )
                })
                .collect(),
            keyspace: keyspace_name.to_string(),
            name: name.to_owned(),
            frozen,
        }
    }
```

This create_with_params function tries to find a user-defined type in keyspace metadata.
If it's not found, the function returns a UDT with empty fields ( `.unwrap_or(&vec![])` ).

But notice that passing these UDT definitions and keyspace name was unnecessary
even before bumping to 1.0.0. The empty fields could be constructed when Err(MissingUserDefinedType)
was present. If the error was present, this meant exactly that UDT name could not be found in keyspace metadata
in the first place on rust-driver side.

The create_with_params function will be addressed/removed later in this PR. Ultimately,
we will also remove get_column_type_from_cql_type().