iterator: implement cass_iterator_type
It still needs to be adjusted, but we need to adjust CassIterator enum hierarchy
first to fix it.

In the end, I want CassIterator enum top-level hierarchy directly correspond to
CassIteratorType enum hierarchy.

For example, notice that we currently don't return `CASS_ITERATOR_TYPE_TUPLE`. This
is because iterator created from tuple is currently hidden under `CassIterator::CassCollectionIterator`.