cargo/cmake: compile with cpp_rust_unstable cfg
Just introducing the config file is unfortunately not enough.
CMakeRust does not seem to pick up the config file.

Even if it did (after some adjustments), it defines the RUSTFLAGS env variable
which takes precedence over the config file. The value of RUSTFLAGS changes
depending on the platform (so we can't just put them in the config file).