row: borrow metadata from result, instead of cloning an Arc
Because of that, we need to introduce a lifetime parameter to CassRow.
CassRow will from now on borrow metadata from CassResult (CassRowsResult).

OTOH, notice that CassRowsResult need to hold the deserialized first row.
This is because `cass_result_first_row` method exists. We don't want to
deserialize the row each time this method is called.

Because of that, the self-borrowing appears in our codebase. We have a
CassRowsResult, which holds a CassRow that borrows CassResultMetadata from the CassRowsResult.
It's not possible to express it in safe Rust, thus we make use of yoke crate
which wraps unsafe code and exposes safe abstractions on top of it.