iterator: introduce CassMapCollectionIterator
This is an iterator obtained from `cass_iterator_from_collection` with map
provided as a value. It wraps `CassMapIterator`. The difference between
`CassMapCollectionIterator` and `CassMapIterator` is that the former
iterates over single values (either key or value) while the latter iterates
over key-value pairs. This is the semantic from cpp-driver.