result: include DeserializedMetadataAndRawRows in yoke cart
Currently `CassValue` wraps CqlValue. This is not true for new CassValue,
which holds serialized bytes. This means, that CassRow will not only borrow
the metadata from result, but also the serialized bytes.

This is why, we need to include `DeserializedMetadataAndRawRows` in the yoke cart
of `RowWithSelfBorrowedResultData`. This costs us additional allocation of Arc.
I think it is a fair exchange - one allocation per result, for a guaranteed API safety.

Notice the changes I did to unit test. We no longer construct any rows (values)
there. This is because we don't use CqlValue in RowWithSelfBorrowedResultData API anymore.
This does not hurt us though, because those values were not used in the test
anyway. We only test result metadata.

Note: Maybe it's a good idea to add the ser/deser unit tests in the future.
There are currently many cpp integration tests for this, but it would be
nice to test the API in Rust.