it: adjust and enable HeartbeatFailed
Since we implemented metrics, we can enable HeartbeatFailed test with some
adjustments to log filtering.

This test seems to fail under valgrind. This is why I enable it to run
next to other test that cannot be run under valgrind.

Note: The original test seems to be flaky for Cassandra. The following scenario
occurred:
1. node2 is paused
2. keepaliver notifies the pool refiller about that
3. refiller removes the connection to node2 (metrics::total_connections -= 1)
4. in the test, we read get_metrics().total_connections < initial_connections - we go out of the loop
5. refiller tries to open a connection again (metrics::total_connections += 1)
6. we read get_metrics().total_connections, and expect total_connections to be
less than initial_connections - but it is not.

This is why, to combat this, I adjusted the test so the same metrics snapshot
is used to leave the loop and make an assertion. In this case, the aforementioned
"unlucky" scenario will not happen.