IT: Adjust and enable "Requests" metrics test
Why does this test not work without adjustments? Well, this is because
rust-driver collects latencies with millisecond graunularity. In result,
most of the latencies during the tests in local setup are rounded to 0ms.

This is why, we somehow need to simulate higher latencies during the test.
There is one piece of code that user controls and is executed in rust-driver
in between start and end time measurements - namely `HistoryListener::log_attempt_start`.

This is where we can add a sleep to simulate higher latencies in local setup.
And so, I implemented `SleepingHistoryListener` that does just that. In addition,
I implemented the testing API to set such listener on the statement.

The "Requests" test is adjusted accordingly, and enabled.

Note: Since all latencies during the test in local setup are now expected to be around 1ms,
I loosened the stddev check to be `>= 0` instead of `> 0`.