retry_policy: implement cass_retry_policy_logging_new
It's used in some tests, that's why we need to implement it.

The semantics of this retry policies is a bit different in C++ driver
and in Rust driver. Therefore, the logs are not exactly the same.
So in order to enable the tests using this retry policy, we will need to
adjust the tests to expect the new format of the logs.