lbp: introduce utility methods for set operations
Motivation is going to be further explained in the later commit,
where I explain how cpp-driver decides to which hosts it opens
the connections. In short: if **all** execution profiles ignore the host, the driver
does not open the connection to it.

Knowing how filtering rules are applied, I noticed that we can implement
the filtering rules for host filter by computing the unions and intersections
of whitelists and blacklists respectively.
Note: Empty list means that this filtering rule is disabled.