it: explicitly disable UDF server-side failure tests
It will be now more clear which tests are failing and should possibly
be addressed in the future. The disabled tests from this suite are using UDFs,
which are hidden behind unstable feature in Scylla. We may want to
enable them in the future, but for now let's leave them disabled.